+++ src/Mapline.Tests/Helpers/MaplineDbContextSettings.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Mapline.Tests.Helpers
+{
+    public class MaplineDbContextSettings
+    {
+        public string DatabaseName { get; set; } = "Mapline";
+        public bool InitializeData { get; set; } = true;
+        public bool CreateModel { get; set; } = false;
+        public string LanguageFolder { get; set; } = ".\\..\\..\\..\\..\\..\\data\\Language";
+    }
+}
