+++ src/Mapline.Tests/WebTests/DataTests/MaplineDbContextTests.cs
@@ -1,4 +1,5 @@
-﻿using Mapline.Web.Data;
+﻿using Mapline.Tests.Helpers;
+using Mapline.Web.Data;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
